# src/main/resources/SwaggerAPI.yaml
swagger: "2.0"
info:
  description: "This is for hr connect support tool server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "HR Connect Process Support Tool"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "saarnab@amazon.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/v1"
tags:
  - name: "APIs"
    description: "We have the following three APIs"
    externalDocs:
      description: "Find out more"
      url: "http://swagger.io"

schemes:
  - "https"
  - "http"

paths:
  /product:
    get:
      tags:
        - "product"
      summary: "Get the list of all Msg groups for the dropdown"
      description: ""
      operationId: "hr-cps-tool-get-1"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "productName"
          description: "Name of the Msg Group"
          required: true
          type: "string"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid Msg Group supplied"
        "404":
          description: "Not found"
        "405":
          description: "Invalid input"

  /component:
    get:
      tags:
        - "component"
      summary: "Get the rows corresponding to the Msg groups in the dropdown"
      description: ""
      operationId: "hr-cps-tool-get-2"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "productName"
          description: "Name of the Msg Group"
          required: true
          type: "string"
        - in: "query"
          name: "componentName"
          description: "Name of the DDB table"
          required: true
          type: "string"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid Msg Group supplied"
        "404":
          description: "Not found"
        "405":
          description: "Invalid input"

    post:
      tags:
        - "component"
      summary: "Updates a row corresponding to the Msg groups in the dropdown"
      description: ""
      operationId: "hr-cps-tool-post-1"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "productName"
          description: "Name of the Msg Group"
          required: true
          type: "string"
        - in: "query"
          name: "componentName"
          description: "Name of the DDB Table"
          required: true
          type: "string"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid Msg Group supplied"
        "404":
          description: "Not found"

externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"